USAGOV-445-modsecurity-change-for-menu-updates: Remove failed attempt to remove rule 200002, bump up SecRequestBodyNoFilesLimit instead.
We need a larger SecRequestBodyNoFilesLimit to accommodate edits to the menu structure; the
REQUEST payloads on these are currently running a little over 200k for the left nav (English).
Unfortunately SecRequestBodyNoFilesLimit cannot be adjusted on a per-request basis.
Rule 200002 is just about what to do after failing to process a request body, so overriding
it wasn't going to help.