USAGOV-762-Mobile-Nav-caching-fixes: Make separate blocks for mobile menus with different caching rules.
The mobile menu is usually a rendering of either the "Left Menu
English" or "Left Menu Spanish" menu. However, for (federal) Directory Records,
the "menu" is a hard-coded structure PLUS the title of the current page. So for
Directory Records, the mobile menu must appear in a block that is based on the
current page and language, not on the left nav menu and language. This enables
Drupal to cache the block appropriately.